Refactor decomposition algorithms
