+++ include/tweedledum/Decomposition/BridgeDecomposer.h
@@ -4,13 +4,24 @@
 *-----------------------------------------------------------------------------*/
 #pragma once
 
-#include "../../IR/Circuit.h"
+#include "../IR/Circuit.h"
+#include "../IR/Instruction.h"
+#include "../Target/Device.h"
 
 #include <nlohmann/json.hpp>
 
 namespace tweedledum {
 
-Circuit barenco_decomp(
-  Circuit const& original, nlohmann::json const& config = {});
+class BridgeDecomposer {
+public:
+    BridgeDecomposer(Device const& device, nlohmann::json const& config = {})
+        : device_(device)
+    {}
+
+    bool decompose(Circuit& circuit, Instruction const& inst);
+
+private:
+    Device const& device_;
+};
 
 } // namespace tweedledum
