+++ src/Passes/Decomposition/bridge_decomp.cpp
@@ -4,53 +4,21 @@
 *-----------------------------------------------------------------------------*/
 #include "tweedledum/Passes/Decomposition/bridge_decomp.h"
 
+#include "tweedledum/Decomposition/BridgeDecomposer.h"
 #include "tweedledum/Operators/Extension/Bridge.h"
 #include "tweedledum/Operators/Standard/X.h"
 #include "tweedledum/Passes/Utility/shallow_duplicate.h"
 
 namespace tweedledum {
 
-namespace {
-
-inline bool decompose(
-  Device const& device, Circuit& circuit, Instruction const& inst)
+Circuit bridge_decomp(
+  Device const& device, Circuit const& original, nlohmann::json const& config)
 {
-    std::vector<uint32_t> const path =
-      device.shortest_path(inst.control(), inst.target());
-    assert(path.size() >= 3u);
-    for (uint32_t i = 1; i < path.size() - 1; ++i) {
-        circuit.apply_operator(
-          Op::X(), {Qubit(path.at(i)), Qubit(path.at(i + 1))});
-    }
-    for (uint32_t i = path.size() - 2; i-- > 1;) {
-        circuit.apply_operator(
-          Op::X(), {Qubit(path.at(i)), Qubit(path.at(i + 1))});
-    }
-    for (uint32_t i = 0; i < path.size() - 1; ++i) {
-        circuit.apply_operator(
-          Op::X(), {Qubit(path.at(i)), Qubit(path.at(i + 1))});
-    }
-    for (uint32_t i = path.size() - 2; i-- > 0;) {
-        circuit.apply_operator(
-          Op::X(), {Qubit(path.at(i)), Qubit(path.at(i + 1))});
-    }
-    return true;
-}
-
-} // namespace
-
-void bridge_decomp(
-  Device const& device, Circuit& circuit, Instruction const& inst)
-{
-    decompose(device, circuit, inst);
-}
-
-Circuit bridge_decomp(Device const& device, Circuit const& original)
-{
+    BridgeDecomposer decomposer(device, config);
     Circuit decomposed = shallow_duplicate(original);
     original.foreach_instruction([&](Instruction const& inst) {
         if (inst.is_a<Op::Bridge>()) {
-            decompose(device, decomposed, inst);
+            decomposer.decompose(decomposed, inst);
             return;
         }
         decomposed.apply_operator(inst);
