+++ src/Passes/Decomposition/parity_decomp.cpp
@@ -4,35 +4,19 @@
 *-----------------------------------------------------------------------------*/
 #include "tweedledum/Passes/Decomposition/parity_decomp.h"
 
-#include "tweedledum/Operators/Extension.h"
-#include "tweedledum/Operators/Standard.h"
+#include "tweedledum/Decomposition/ParityDecomposer.h"
+#include "tweedledum/Operators/Extension/Parity.h"
 #include "tweedledum/Passes/Utility/shallow_duplicate.h"
 
 namespace tweedledum {
 
-namespace {
-
-inline bool decompose(Circuit& circuit, Instruction const& inst)
+Circuit parity_decomp(Circuit const& original, nlohmann::json const& config)
 {
-    inst.foreach_control([&](Qubit qubit) {
-        circuit.apply_operator(Op::X(), {qubit, inst.target()}, inst.cbits());
-    });
-    return true;
-}
-
-} // namespace
-
-void parity_decomp(Circuit& circuit, Instruction const& inst)
-{
-    decompose(circuit, inst);
-}
-
-Circuit parity_decomp(Circuit const& original)
-{
+    ParityDecomposer decomposer(config);
     Circuit decomposed = shallow_duplicate(original);
     original.foreach_instruction([&](Instruction const& inst) {
         if (inst.is_a<Op::Parity>()) {
-            decompose(decomposed, inst);
+            decomposer.decompose(decomposed, inst);
             return;
         }
         decomposed.apply_operator(inst);
